requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0  # for handling environment variables
Pillow>=10.0.0  # For PIL/Image processing
piexif>=1.1.3   # For EXIF metadata extraction
moviepy>=1.0.3
requests>=2.31.0 